refactor: MIN, MAX COUNT 상수 한 파일에서 처리
