+++ frontend/src/constants/cart.ts
@@ -1,0 +1,2 @@
+export const MAX_COUNT = 49
+export const MIN_COUNT = 1
